+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.40.0] - 2023-11-27
+
+### Added
+
+### Fixed
+
+### Changed
+
+- [change](https://github.com/FrankC01/pysui/issues/173) Removed constraint on SuiTransaction.make_move_vector that failed if arguments were pure.
+- TransactionData added 3 new transaction kinds: AuthenticatorStateUpdate, RandomnessStateUpdate and EndOfEpochTransaction.
+
+### Removed
+
 ## [0.39.1] - Unpublished
 
 ### Added